Named Entity Recognition on multiple languages
In this notebook, we took 4 languages and performed multilingual NER comparing 2 approaches of training on the single language at a time and training on all the concatenated languages. The result shows when trained on all languages combined give significant boost as compared to the other scenario. This boost however requires that the languages in the context are from same family.