text generation using gpt2
In this notebook, we see how to generate model output using a gpt2 model. We tried gpt2-xl but it was quite heavy to load on local as well as the colab environment. 
We discuss greedy decoding method and beam search decoding method followed by sampling methods with temperature control and top-k and top-p twists.